Replace buildPresetLabels func
