+++ client/src/app/components/Icons/IconedStatus.tsx
@@ -13,7 +13,6 @@ import {
   UnknownIcon,
   TopologyIcon,
 } from "@patternfly/react-icons";
-import { buildPresetLabels } from "@app/pages/applications/components/application-analysis-status";
 
 export type IconedStatusPreset =
   | "InheritedReviews"
@@ -164,3 +163,22 @@ export const IconedStatus: React.FC<IIconedStatusProps> = ({
     />
   );
 };
+
+export const buildPresetLabels = (
+  t: (key: string) => string
+): Record<IconedStatusPreset, { label: string }> => ({
+  InProgressInheritedReviews: { label: t("terms.inProgress") },
+  InProgressInheritedAssessments: { label: t("terms.inProgress") },
+  InheritedReviews: { label: t("terms.completed") },
+  InheritedAssessments: { label: t("terms.completed") },
+  Canceled: { label: t("terms.canceled") },
+  Completed: { label: t("terms.completed") },
+  CompletedWithErrors: { label: t("terms.completedWithErrors") },
+  Error: { label: t("terms.error") },
+  Failed: { label: t("terms.failed") },
+  InProgress: { label: t("terms.inProgress") },
+  NotStarted: { label: t("terms.notStarted") },
+  Scheduled: { label: t("terms.scheduled") },
+  Ok: { label: t("terms.ok") }, // Add Ok with a label
+  Unknown: { label: t("terms.unknown") }, // Add Unknown with a label
+});
